move upgrade test to upgrade test file
